Consistently order tests and code
